requirements.txt
Pillow==9.5.0
URLObject==2.4.3
requests
